third_party/blink/web_tests/editing/execCommand/remove-format-image-expected.txt

This tests RemoveFormant command not removing an image.  You should see one image element in each of the test results

An image between text:
| "<#selection-anchor>hello"
| " "
| <img>
|   src="../resources/abe.png"
| " world<#selection-focus>"

RemoveFormat on an image followed by JustifyNone and FormatBlock with p.:
| <p>
|   <#selection-anchor>
|   <img>
|     src="../resources/abe.png"
|   <#selection-focus>
